pyproject.toml
[project]
name = "lgdl-mvp"
version = "0.1.0"
description = "LGDL MVP: Lark grammar + runtime with optional OpenAI embeddings"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Your Team" }]

# Runtime dependencies (all offline-friendly)
dependencies = [
  "fastapi>=0.112,<1",
  "uvicorn[standard]>=0.30,<1",
  "pydantic>=2.7,<3",
  "lark>=1.1,<2",
  "jsonschema>=4.22,<5",
  "click>=8.1,<9",
  "numpy>=1.26,<3"
]

[project.optional-dependencies]
# Installs the HTTP client the matcher uses for OpenAI embeddings
openai = [
  "httpx>=0.27,<1"
]

# Test & lint extras
dev = [
  "pytest>=8,<9",
  "pytest-asyncio>=0.23,<1",
  "requests>=2.31,<3",
  "PyYAML>=6,<7",
  "ruff>=0.6,<1"
]

[project.scripts]
# CLI entrypoint (so you can: uv run lgdl validate examples/medical/game.lgdl)
lgdl = "lgdl.cli.main:cli"

[tool.uv]
package = true


[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-q"

# (Optional) ruff config if you enable it
[tool.ruff]
line-length = 100
target-version = "py310"
